added script to initialize and populate MYSQL database

+++ init/dbstartup.sql
@@ -13,7 +13,7 @@ INSERT INTO genres (genre_id, name) VALUES (5, "Musical");
 
 
 
-CREATE TABLE movies (id BIGINT NOT NULL, title VARCHAR(1000) NOT NULL, year BIGINT, director VARCHAR(1000), revenue BIGINT, rating FLOAT, genre BIGINT, PRIMARY KEY (id));
+CREATE TABLE movies (id BIGINT NOT NULL, title VARCHAR(1000) NOT NULL, year BIGINT, director VARCHAR(1000), revenue BIGINT, rating DECIMAL, genre BIGINT, PRIMARY KEY (id));
 
 INSERT INTO movies (id, title, year, director, revenue, rating, genre) VALUES (101, "Gone with the Wind", 1939, "Victor Fleming", 87432102, 4.1, 2);
 INSERT INTO movies (id, title, year, director, revenue, rating, genre) VALUES (102, "Star Wars", 1977, "Robert Wise", 143192342, 3.9, 1);
